requirement.txt
streamlit
torch
torchvision
grad-cam
opencv-python-headless
pillow
numpy